Another warning fix

+++ libretro-common/formats/xml/rxml.c
@@ -100,7 +100,7 @@ rxml_document_t *rxml_load_document(const char *path)
       return NULL;
 
    len                     = filestream_get_size(file);
-   memory_buffer           = (char*)malloc(len + 1);
+   memory_buffer           = (char*)malloc((size_t)(len + 1));
    if (!memory_buffer)
       goto error;
